+++ hazelcast/src/main/resources/META-INF/services/com.hazelcast.DataSerializerHook
@@ -1,6 +1,6 @@
 com.hazelcast.internal.cluster.impl.ClusterDataSerializerHook
 com.hazelcast.spi.impl.SpiDataSerializerHook
-com.hazelcast.partition.PartitionDataSerializerHook
+com.hazelcast.internal.partition.impl.PartitionDataSerializerHook
 com.hazelcast.map.impl.MapDataSerializerHook
 com.hazelcast.collection.impl.queue.QueueDataSerializerHook
 com.hazelcast.multimap.impl.MultiMapDataSerializerHook
