+++ hazelcast/src/main/resources/META-INF/services/com.hazelcast.PortableHook
@@ -1,5 +1,5 @@
 com.hazelcast.spi.impl.SpiPortableHook
-com.hazelcast.partition.PartitionPortableHook
+com.hazelcast.internal.partition.impl.PartitionPortableHook
 com.hazelcast.mapreduce.impl.MapReducePortableHook
 com.hazelcast.replicatedmap.impl.client.ReplicatedMapPortableHook
 com.hazelcast.ringbuffer.impl.client.RingbufferPortableHook
